relocate tests
