+++ tests/bofire/surrogates/test_gps.py
@@ -19,7 +19,6 @@ from bofire.benchmarks.api import Himmelblau
 from bofire.data_models.domain.api import Inputs, Outputs
 from bofire.data_models.enum import CategoricalEncodingEnum, RegressionMetricsEnum
 from bofire.data_models.features.api import (
-    CategoricalDescriptorInput,
     CategoricalInput,
     ContinuousInput,
     ContinuousOutput,
@@ -32,9 +31,6 @@ from bofire.data_models.kernels.api import (
     ScaleKernel,
 )
 from bofire.data_models.molfeatures.api import (
-    Fingerprints,
-    FingerprintsFragments,
-    Fragments,
     MordredDescriptors,
 )
 from bofire.data_models.priors.api import (
@@ -52,322 +48,8 @@ from bofire.data_models.surrogates.api import (
     SingleTaskGPSurrogate,
 )
 from bofire.data_models.surrogates.trainable import metrics2objectives
-from bofire.surrogates.utils import (
-    get_categorical_feature_keys,
-    get_continuous_feature_keys,
-    get_molecular_feature_keys,
-    get_scaler,
-)
-from bofire.utils.torch_tools import tkwargs
 
 RDKIT_AVAILABLE = importlib.util.find_spec("rdkit") is not None
-
-
-@pytest.mark.parametrize(
-    "scaler_enum, input_preprocessing_specs, expected_scaler, expected_indices, expected_offset, expected_coefficient",
-    [
-        (
-            ScalerEnum.NORMALIZE,
-            {
-                "x_cat": CategoricalEncodingEnum.ONE_HOT,
-                "x_desc": CategoricalEncodingEnum.ONE_HOT,
-            },
-            Normalize,
-            torch.tensor([0, 1], dtype=torch.int64),
-            torch.tensor([-4.0, -4.0]).to(**tkwargs),
-            torch.tensor([8.0, 8.0]).to(**tkwargs),
-        ),
-        (
-            ScalerEnum.NORMALIZE,
-            {
-                "x_cat": CategoricalEncodingEnum.ONE_HOT,
-                "x_desc": CategoricalEncodingEnum.DESCRIPTOR,
-            },
-            Normalize,
-            torch.tensor([0, 1, 2], dtype=torch.int64),
-            torch.tensor([-4.0, -4.0, 1.0]).to(**tkwargs),
-            torch.tensor([8.0, 8.0, 5.0]).to(**tkwargs),
-        ),
-        (
-            ScalerEnum.STANDARDIZE,
-            {
-                "x_cat": CategoricalEncodingEnum.ONE_HOT,
-                "x_desc": CategoricalEncodingEnum.ONE_HOT,
-            },
-            InputStandardize,
-            torch.tensor([0, 1], dtype=torch.int64),
-            None,
-            None,
-        ),
-        (
-            ScalerEnum.STANDARDIZE,
-            {
-                "x_cat": CategoricalEncodingEnum.ONE_HOT,
-                "x_desc": CategoricalEncodingEnum.DESCRIPTOR,
-            },
-            InputStandardize,
-            torch.tensor([0, 1, 2], dtype=torch.int64),
-            None,
-            None,
-        ),
-        (
-            ScalerEnum.IDENTITY,
-            {
-                "x_cat": CategoricalEncodingEnum.ONE_HOT,
-                "x_desc": CategoricalEncodingEnum.ONE_HOT,
-            },
-            type(None),
-            None,
-            None,
-            None,
-        ),
-        (
-            ScalerEnum.IDENTITY,
-            {
-                "x_cat": CategoricalEncodingEnum.ONE_HOT,
-                "x_desc": CategoricalEncodingEnum.DESCRIPTOR,
-            },
-            type(None),
-            None,
-            None,
-            None,
-        ),
-    ],
-)
-def test_get_scaler(
-    scaler_enum,
-    input_preprocessing_specs,
-    expected_scaler,
-    expected_indices,
-    expected_offset,
-    expected_coefficient,
-):
-    inputs = Inputs(
-        features=[
-            ContinuousInput(
-                key=f"x_{i+1}",
-                bounds=(-4, 4),
-            )
-            for i in range(2)
-        ]
-        + [
-            CategoricalInput(key="x_cat", categories=["mama", "papa"]),
-            CategoricalDescriptorInput(
-                key="x_desc",
-                categories=["alpha", "beta"],
-                descriptors=["oskar"],
-                values=[[1], [6]],
-            ),
-        ]
-    )
-    experiments = inputs.sample(n=10)
-    scaler = get_scaler(
-        inputs=inputs,
-        input_preprocessing_specs=input_preprocessing_specs,
-        scaler=scaler_enum,
-        X=experiments[inputs.get_keys()],
-    )
-    assert isinstance(scaler, expected_scaler)
-    if expected_indices is not None:
-        assert (scaler.indices == expected_indices).all()
-    else:
-        with pytest.raises(AttributeError):
-            assert (scaler.indices == expected_indices).all()
-    if expected_offset is not None:
-        assert torch.allclose(scaler.offset, expected_offset)
-        assert torch.allclose(scaler.coefficient, expected_coefficient)
-    else:
-        if scaler is None:
-            with pytest.raises(AttributeError):
-                assert (scaler.offset == expected_offset).all()
-            with pytest.raises(AttributeError):
-                assert (scaler.coefficient == expected_coefficient).all()
-
-
-@pytest.mark.parametrize(
-    "scaler_enum, input_preprocessing_specs, expected_scaler, expected_indices_length",
-    [
-        (
-            ScalerEnum.NORMALIZE,
-            {
-                "x_mol": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
-            },
-            Normalize,
-            4,
-        ),
-        (
-            ScalerEnum.NORMALIZE,
-            {
-                "x_mol": Fingerprints(n_bits=32),
-            },
-            Normalize,
-            2,
-        ),
-        (
-            ScalerEnum.STANDARDIZE,
-            {
-                "x_mol": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
-            },
-            InputStandardize,
-            4,
-        ),
-        (
-            ScalerEnum.STANDARDIZE,
-            {
-                "x_mol": Fragments(),
-            },
-            InputStandardize,
-            2,
-        ),
-        (
-            ScalerEnum.IDENTITY,
-            {
-                "x_mol": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
-            },
-            type(None),
-            0,
-        ),
-        (
-            ScalerEnum.IDENTITY,
-            {
-                "x_mol": FingerprintsFragments(n_bits=32),
-            },
-            type(None),
-            0,
-        ),
-    ],
-)
-def test_get_scaler_molecular(
-    scaler_enum,
-    input_preprocessing_specs,
-    expected_scaler,
-    expected_indices_length,
-):
-    inputs = Inputs(
-        features=[
-            ContinuousInput(
-                key=f"x_{i+1}",
-                bounds=(0, 5),
-            )
-            for i in range(2)
-        ]
-        + [MolecularInput(key="x_mol")]
-    )
-    experiments = [
-        [5.0, 2.5, "CC(=O)Oc1ccccc1C(=O)O"],
-        [4.0, 2.0, "c1ccccc1"],
-        [3.0, 0.5, "[CH3][CH2][OH]"],
-        [1.5, 4.5, "N[C@](C)(F)C(=O)O"],
-    ]
-    experiments = pd.DataFrame(experiments, columns=["x_1", "x_2", "x_mol"])
-    scaler = get_scaler(
-        inputs=inputs,
-        input_preprocessing_specs=input_preprocessing_specs,
-        scaler=scaler_enum,
-        X=experiments[inputs.get_keys()],
-    )
-    assert isinstance(scaler, expected_scaler)
-    if expected_indices_length != 0:
-        assert len(scaler.indices) == expected_indices_length
-    else:
-        with pytest.raises(
-            AttributeError, match="'NoneType' object has no attribute 'indices'"
-        ):
-            assert scaler.indices is None
-
-
-@pytest.mark.skipif(not RDKIT_AVAILABLE, reason="requires rdkit")
-@pytest.mark.parametrize(
-    "specs, expected_molecular_keys, expected_continuous_keys, expected_categorical_keys",
-    [
-        (
-            {
-                "x2": CategoricalEncodingEnum.ONE_HOT,
-                "x3": CategoricalEncodingEnum.ONE_HOT,
-                "x4": Fingerprints(n_bits=2),
-            },
-            ["x4"],
-            ["x1"],
-            ["x2", "x3"],
-        ),
-        (
-            {
-                "x2": CategoricalEncodingEnum.ONE_HOT,
-                "x3": CategoricalEncodingEnum.ONE_HOT,
-                "x4": Fragments(fragments=["fr_unbrch_alkane", "fr_thiocyan"]),
-            },
-            ["x4"],
-            ["x1"],
-            ["x2", "x3"],
-        ),
-        (
-            {
-                "x2": CategoricalEncodingEnum.ONE_HOT,
-                "x3": CategoricalEncodingEnum.ONE_HOT,
-                "x4": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
-            },
-            [],
-            ["x4", "x1"],
-            ["x2", "x3"],
-        ),
-        (
-            {
-                "x2": CategoricalEncodingEnum.ONE_HOT,
-                "x3": CategoricalEncodingEnum.DESCRIPTOR,
-                "x4": Fingerprints(n_bits=2),
-            },
-            ["x4"],
-            ["x1", "x3"],
-            ["x2"],
-        ),
-        (
-            {
-                "x2": CategoricalEncodingEnum.ONE_HOT,
-                "x3": CategoricalEncodingEnum.DESCRIPTOR,
-                "x4": Fragments(fragments=["fr_unbrch_alkane", "fr_thiocyan"]),
-            },
-            ["x4"],
-            ["x1", "x3"],
-            ["x2"],
-        ),
-        (
-            {
-                "x2": CategoricalEncodingEnum.ONE_HOT,
-                "x3": CategoricalEncodingEnum.DESCRIPTOR,
-                "x4": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
-            },
-            [],
-            ["x4", "x1", "x3"],
-            ["x2"],
-        ),
-    ],
-)
-def test_get_feature_keys(
-    specs,
-    expected_molecular_keys,
-    expected_continuous_keys,
-    expected_categorical_keys,
-):
-    inps = Inputs(
-        features=[
-            ContinuousInput(key="x1", bounds=(0, 1)),
-            CategoricalInput(key="x2", categories=["apple", "banana", "orange"]),
-            CategoricalDescriptorInput(
-                key="x3",
-                categories=["apple", "banana", "orange", "cherry"],
-                descriptors=["d1", "d2"],
-                values=[[1, 2], [3, 4], [5, 6], [7, 8]],
-            ),
-            MolecularInput(key="x4"),
-        ]
-    )
-    molecular_feature_keys = get_molecular_feature_keys(specs)
-    continuous_feature_keys = get_continuous_feature_keys(inps, specs)
-    categorical_feature_keys = get_categorical_feature_keys(specs)
-
-    assert molecular_feature_keys == expected_molecular_keys
-    assert continuous_feature_keys == expected_continuous_keys
-    assert categorical_feature_keys == expected_categorical_keys
 
 
 @pytest.mark.parametrize(
